Fix unreliable input signal
Previously the master volume was controlled by controlling the volume
of the input signal. This was bad because the envelope followers /
gates in the snare synthesis couldn't rely on a consistent signal level.
By setting the master volume in the output mixer, this issue is solved.

This is only possible with a new walkman version, where the amplitude
of a mixer module can be modified.