Modernize dependencies: replace buildout with nix
In further development of walkman, buildout was dropped in favour of
nix, that's used now. By updating to nix, we also update walkman as the
old version we used with buildout isn't supported by 'mutwo-nix'.
Because this is a newer python version now, the sys command doesn't
exist anymore, so let's remove it and hope it wasn't so important in
improving the performance of the program.